feat(entry): manually added entry support
This implements a basic read-it-later system that allows users to manually add
entries to the database, and a new option on feeds that controls if they are
for these new manual entries.

Feature support in this commit is somewhat bare-bones.
Following commits are intended to further flesh out read-it-later functionality.

Much inspiration was taken from the Shiori project for this, my initial goal was
to replace the combination of Miniflux and Shiori with just Miniflux.

Closes: #329 https://github.com/miniflux/v2/issues/329
See-also: #3719 https://github.com/orgs/miniflux/discussions/3719