feat(feeds): Uncategorized feeds support
Adds support for the category_id of feeds to be NULL.
This required a number of code changes throughout the application, making this
a fairly substantial change. I hoped to get it done all in one go.

When the category_id or Category field of a feed is NULL it is considered part
of the "Uncategorized" pseudo-category.