refactor(entry): Unified tag formatting
There was duplication and inconsistency in the way tags were parsed
which could lead to unexpected behavior for users.
This change unifies onto model.CleanTags for all tag parsing.

Tags are also now converted to lowercase when importing. This is to improve
deduplication. Existing tags will continue to match new lowercased tags since
the database query was already case-insensitive.