feat(entry): Basic Entry editing UI
This implements a simple UI for editing entries.
It is primarily useful for manually added entries, but can be used for feed
provided entries as well.

Alongside manually added entries in the previous commit this completes the
CRUD operations for entries.

Notably this exposes the Entry tagging feature to the user more prominently
so I am marking the related task as closed.

Closes: https://github.com/miniflux/v2/issues/253
See-also: https://github.com/orgs/miniflux/discussions/3719